+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - **Fix:** Since 2.2, team members are unable to post new resource updates (#5)
 - **Fix:** User who is not team member and made the release cannot download without permission (#7)
 - **Fix:** Username/user id columns added in resource update/version are not prefixed (#8)
+- **Fix:** Since 2.2, Resource update thread reply is made by resource owner (#9)
 
 ## 1.1.0 (`1000070`)